+++ firmware/hdl/packages/gtl_pkg.vhd
@@ -70,6 +70,7 @@ constant TC_COND_STAGES: natural := 2; -- pipeline stages for "Towercount condit
 constant EXT_COND_STAGES: natural := 2; -- pipeline stages for "External conditions" to get same pipeline to algos as conditions
 constant CENTRALITY_STAGES: natural := 2; -- pipeline stages for "Centrality" to get same pipeline to algos as conditions
 constant MUS_STAGES: natural := 2; -- pipeline stages for "Hadronic shower triggers (muon)" to get same pipeline to algos as conditions
+constant ADT_SIM_DEL: natural := 2; -- delay of ADT for simulation 
 constant INTERMEDIATE_PIPELINE: boolean := true; -- intermediate pipeline
 constant CONDITIONS_PIPELINE: boolean := true; -- pipeline at output of conditions