+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+<<<<<<< HEAD
 ## [v1.22.0] - 2023-03-06
 ### Comment
 
@@ -29,6 +30,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - ../payload/gtl/muon_obj_cuts.vhd
   - ../payload/gtl/comb_conditions.vhd
   - ../payload/gtl/correlation_conditions.vhd
+=======
+## [v1.21.8] - 2022-11-29
+### Comment
+
+- mp7_ugt firmware release v1.21.8
+- test version for "anomaly detection trigger (ADT)"
+
+### Changed
+- source files:
+  - ../payload/gtl_module_tpl.vhd
+>>>>>>> anomaly_detection_trigger
 
 ## [v1.21.7] - 2022-11-29
 ### Comment
@@ -95,9 +107,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - source files:
   - ../packages/gt_mp7_core_pkg.vhd
 
-### Added
-- source files:
-  - ../packages/gt_mp7_top_pkg_sim.vhd
+
 
 ## [v1.21.2] - 2022-09-28
 ### Comment
@@ -121,6 +131,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - source files:
   - ../packages/gt_mp7_core_pkg.vhd
+>>>>>>> master
 
 ## [v1.21.0] - 2022-09-14
 ### Comment